Allow nodeunit:webserver to live reload the specified tests
Tests to run specified in karma.conf.js are now reloaded live from the web server which helps with testing
